+++ tests/test_visitors/test_ast/test_general/test_comparison_order_func.py
@@ -2,9 +2,6 @@
 
 import pytest
 
-from wemake_python_styleguide.violations.consistency import (
-    ComparisonOrderViolation,
-)
 from wemake_python_styleguide.visitors.ast.order import WrongOrderVisitor
 
 # Templates to be checked
@@ -29,6 +26,12 @@ if index > some_object.get_index():
     return 1
 """
 
+more_than_one_variable = """
+if 0 < b < c:
+    return 1
+"""
+
+
 @pytest.mark.parametrize('code', [
     if_with_func_call,
     if_with_complex_call_1,
@@ -36,6 +39,24 @@ if index > some_object.get_index():
     if_with_method_call,
 ])
 def test_functions_methods_comparison(
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+):
+    """Testing : no violations with correct comparisons."""
+    tree = parse_ast_tree(code)
+
+    visitor = WrongOrderVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize('code', [
+    more_than_one_variable,
+])
+def test_more_variables_comparison(
     assert_errors,
     parse_ast_tree,
     code,
